Rename footer and update markdown css
