+++ components/posts/post-body.js
@@ -4,8 +4,7 @@ export default function PostBody({ content }) {
   return (
     <div className="max-w-2xl mx-auto">
       <div
-        // className={markdownStyles['markdown']}
-        className={markdownStyles['prose']}
+        className={markdownStyles['markdown']}
         dangerouslySetInnerHTML={{ __html: content }}
       />
     </div>
