+++ components/posts/post-layout.js
@@ -3,7 +3,7 @@ export default function PostLayout({children}) {
     <article>
       <div className="text-base max-w-none prose prose-lg pb-6 lg:pb-16">
         {children}
-        </div>
+      </div>
     </article>
   )
 }
